+++ app/src/main/java/com/example/nba/ui/players_list/PlayerItem.kt
@@ -1,18 +1,14 @@
 package com.example.nba.ui.players_list
 
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -27,11 +23,11 @@ import com.example.nba.ui.navigation.NBAActivity.Companion.PREVIEW_DEFAULT_PLAYE
 @Preview(showBackground = true)
 @OptIn(ExperimentalGlideComposeApi::class)
 @Composable
-inline fun PlayerItem(
+fun PlayerItem(
     modifier: Modifier = Modifier,
     playerIO: PlayerIO = PREVIEW_DEFAULT_PLAYER,
-    crossinline onItemClicked: (playerId: Long?) -> Unit = {},
-    crossinline onTeamClicked: (teamId: Long?) -> Unit = {},
+    onItemClicked: (playerId: Long?) -> Unit = {},
+    onTeamClicked: (teamId: Long?) -> Unit = {},
 ) {
     ConstraintLayout(
         modifier
@@ -45,7 +41,7 @@ inline fun PlayerItem(
             }
             .padding(horizontal = 12.dp, vertical = 12.dp)
     ) {
-        val (txtName, txtHeight, txtWeight, txtPosition, team, imgWeight, imgHeight) = createRefs()
+        val (txtName, txtHeight, txtWeight, txtPosition, imgWeight, imgHeight) = createRefs()
 
         Text(
             text = (playerIO.firstName?: "") + " " + (playerIO.lastName ?: ""),
@@ -57,36 +53,7 @@ inline fun PlayerItem(
             }
         )
         if(playerIO.team != null) {
-            Column(
-                modifier = Modifier
-                    .padding(top = 4.dp)
-                    .clip(RoundedCornerShape(8.dp))
-                    .border(0.5.dp, MaterialTheme.colorScheme.tertiary, RoundedCornerShape(8.dp))
-                    .clickable(
-                        remember { MutableInteractionSource() },
-                        indication = rememberRipple(color = MaterialTheme.colorScheme.surfaceTint),
-                        enabled = true
-                    ) {
-                        onTeamClicked.invoke(playerIO.team.id)
-                    }.constrainAs(team) {
-                        end.linkTo(parent.end)
-                        top.linkTo(parent.top)
-                    }.padding(horizontal = 8.dp, vertical = 4.dp),
-                verticalArrangement = Arrangement.Top,
-                horizontalAlignment = Alignment.End
-            ) {
-                Text(
-                    text = playerIO.team.name.toString(),
-                    fontSize = 18.sp,
-                    color = MaterialTheme.colorScheme.onPrimary
-                )
-                Text(
-                    text = playerIO.team.city.toString(),
-                    fontSize = 16.sp,
-                    color = MaterialTheme.colorScheme.inversePrimary,
-                    modifier = Modifier.padding(top = 4.dp)
-                )
-            }
+            TeamColumn(playerIO = playerIO, onTeamClicked = onTeamClicked)
         }
         if(playerIO.position.isNullOrBlank().not()) {
             Text(
